+++ contracts/interfaces/IsCLAM.sol
@@ -1,16 +1,18 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 pragma solidity >=0.7.5;
 
-import "./IERC20.sol";
+import './IERC20.sol';
 
 interface IsCLAM is IERC20 {
-    function rebase( uint256 clamProfit_, uint epoch_) external returns (uint256);
+    function rebase(uint256 clamProfit_, uint256 epoch_)
+        external
+        returns (uint256);
 
     function circulatingSupply() external view returns (uint256);
 
-    function gonsForBalance( uint amount ) external view returns ( uint );
+    function gonsForBalance(uint256 amount) external view returns (uint256);
 
-    function balanceForGons( uint gons ) external view returns ( uint );
+    function balanceForGons(uint256 gons) external view returns (uint256);
 
-    function index() external view returns ( uint );
+    function index() external view returns (uint256);
 }
